refactor: 모드 아이콘 매핑을 shared로 이동 및 빌드 오류 수정 (#46)
- modeIcons 매핑 테이블을 pages/home/constants에서 shared/components/icons/modes로
  이동해 shared가 pages에 의존하던 역방향 의존성 제거 (ModeIconView가 shared 내부만 참조)
- MeetingIcon의 SVG 속성을 kebab-case(stroke-width)에서 camelCase(strokeWidth)로
  수정해 React 경고 제거
- ModeIconComponentTypes를 ComponentType에서 FunctionComponent로 좁혀
  ModeIconView의 컴포넌트 직접 호출 시 발생하던 tsc 빌드 오류(TS2349) 해결
